requirements.txt
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
Deprecated==1.2.14
exceptiongroup==1.2.0
idna==3.6
iniconfig==2.0.0
packaging==23.2
-e git+https://github.com/tjwaterman99/etl-demo@8ee0658f99506949a421827570f9e26bdc6a3ec2#egg=pipeline
pluggy==1.4.0
psycopg2-binary==2.9.9
pycparser==2.21
PyGithub==2.2.0
PyJWT==2.8.0
PyNaCl==1.5.0
pytest==8.0.2
python-dotenv==1.0.1
requests==2.31.0
-e git+https://github.com/tjwaterman99/etl-demo@8ee0658f99506949a421827570f9e26bdc6a3ec2#egg=theydo_pipeline
tomli==2.0.1
typing_extensions==4.10.0
urllib3==2.2.1
wrapt==1.16.0
